refactor: extract targeting utility
